Solution to Factorial Range

+++ Factorial Range@@67de00956ac3fec9afb90e10/code.c
@@ -1,7 +1,6 @@
 #include <stdio.h>
 
-// Function to compute factorial of a number
-long long factorial(int n) {
+int long factorial(int n) {
     long long fact = 1;
     for (int i = 1; i <= n; i++) {
         fact *= i;
